refactor(events): reorganize event-related API actions and loaders into separate files
- Split monolithic actions.ts and loaders.ts into individual files for each operation
- Update imports across the application to reference new file locations
- Add proper type safety and error handling to all API operations
- Remove unused code and comments from original files
- Ensure consistent naming conventions for API files
