+++ app/presentation/events/routes/update.tsx
@@ -2,8 +2,8 @@ import { PageHeader } from "@/components/common/page-header";
 import { useLoaderData } from "react-router";
 import type { EventEntityWithEventForm } from "~/domain/entities/event.entity";
 import type { LoaderData } from "~/shared/types";
-import { updateEventAction } from "../api/actions";
-import { getEventByIdLoader } from "../api/loaders";
+import { updateEventAction } from "../api/actions/update-event.api-action";
+import { getEventByIdLoader } from "../api/loaders/get-event-by-id.loader";
 import { EventForm } from "../components/event-form";
 
 export const loader = getEventByIdLoader;
